+++ changelog.md
@@ -4,6 +4,7 @@ GDB+ 6.4.3:
 * Update heap analysis functions for pwndbg 
 * Support waitpid(-1) before process traced
 * Allow to setup ptrace_emulation without call to ptrace
+* make Debugger.canary writable 
 GDB+ 6.4.2:
 * handle NOPTRACE for emulate_ptrace and split_on_fork
 * NOP instructions